move token management to basic; fix keep adding secret vol to spec
