+++ drivers/storage/portworx/component/portworx_basic.go
@@ -2,9 +2,14 @@ package component
 
 import (
 	"context"
+	"fmt"
+	"os"
+	"time"
 
 	"github.com/hashicorp/go-version"
+	coreops "github.com/portworx/sched-ops/k8s/core"
 	"github.com/sirupsen/logrus"
+	authv1 "k8s.io/api/authentication/v1"
 	v1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
@@ -13,6 +18,7 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/client-go/tools/record"
+	"k8s.io/kubernetes/pkg/apis/core"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	pxutil "github.com/libopenstorage/operator/drivers/storage/portworx/util"
@@ -33,7 +39,11 @@ const (
 	PxRoleName = "portworx"
 	// PxRoleBindingName name of the Portworx role binding
 	PxRoleBindingName = "portworx"
+	// TokenRefreshTimeKey time to refresh the service account token
+	TokenRefreshTimeKey = "tokenRefreshTime"
 )
+
+var tokenExpirationSeconds = int64(12 * 60 * 60)
 
 type portworxBasic struct {
 	k8sClient client.Client
@@ -84,6 +94,9 @@ func (c *portworxBasic) Reconcile(cluster *corev1.StorageCluster) error {
 	if err := c.createPortworxService(cluster, ownerRef); err != nil {
 		return NewError(ErrCritical, err)
 	}
+	if err := c.createAndMaintainPxSaTokenSecret(cluster, ownerRef); err != nil {
+		return NewError(ErrCritical, err)
+	}
 	if cluster.Spec.Kvdb != nil && cluster.Spec.Kvdb.Internal {
 		if err := c.createPortworxKVDBService(cluster, ownerRef); err != nil {
 			// This should not block portworx installation
@@ -108,6 +121,9 @@ func (c *portworxBasic) Delete(cluster *corev1.StorageCluster) error {
 		return err
 	}
 	if err := k8sutil.DeleteService(c.k8sClient, pxutil.PortworxKVDBServiceName, cluster.Namespace, *ownerRef); err != nil {
+		return err
+	}
+	if err := k8sutil.DeleteSecret(c.k8sClient, pxutil.PortworxServiceAccountTokenSecretName, cluster.Namespace, *ownerRef); err != nil {
 		return err
 	}
 
@@ -519,6 +535,118 @@ func getPortworxServiceSpec(
 	}
 
 	return newService
+}
+
+func (c *portworxBasic) createAndMaintainPxSaTokenSecret(cluster *corev1.StorageCluster, ownerRef *metav1.OwnerReference) error {
+	if err := c.createTokenSecretIfNotExist(cluster, ownerRef); err != nil {
+		return err
+	}
+	if err := c.maintainTokenSecret(cluster, ownerRef); err != nil {
+		return fmt.Errorf("failed to maintain the token secret for px container. %v", err)
+	}
+	return nil
+}
+
+func (c *portworxBasic) createTokenSecretIfNotExist(cluster *corev1.StorageCluster, ownerRef *metav1.OwnerReference) error {
+	secret := &v1.Secret{}
+	err := c.k8sClient.Get(context.TODO(),
+		types.NamespacedName{
+			Name:      pxutil.PortworxServiceAccountTokenSecretName,
+			Namespace: cluster.Namespace,
+		}, secret)
+	if err != nil {
+		if errors.IsNotFound(err) || len(secret.Data) == 0 || len(secret.Data[v1.ServiceAccountTokenKey]) == 0 {
+			token, err := generateToken(cluster)
+			if err != nil {
+				return err
+			}
+			rootCaCrt, _ := os.ReadFile("/var/run/secrets/kubernetes.io/serviceaccount/ca.crt")
+			curTime := time.Now()
+			tokenRefreshTimeBytes, err := curTime.Add(time.Duration(tokenExpirationSeconds / 2)).MarshalBinary()
+			if err != nil {
+				return fmt.Errorf("error marshalling current time to bytes. %v", err)
+			}
+			secret := &v1.Secret{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:            pxutil.PortworxServiceAccountTokenSecretName,
+					Namespace:       cluster.Namespace,
+					OwnerReferences: []metav1.OwnerReference{*ownerRef},
+				},
+				Type: v1.SecretTypeOpaque,
+				Data: map[string][]byte{
+					core.ServiceAccountTokenKey:     token,
+					core.ServiceAccountRootCAKey:    rootCaCrt,
+					core.ServiceAccountNamespaceKey: []byte(cluster.Namespace),
+					TokenRefreshTimeKey:             tokenRefreshTimeBytes,
+				},
+			}
+			if err := k8sutil.CreateOrUpdateSecret(c.k8sClient, secret, ownerRef); err != nil {
+				return err
+			}
+		} else {
+			return fmt.Errorf("error getting the token secret for px from k8s. %v", err)
+		}
+	}
+	return nil
+}
+
+func (c *portworxBasic) maintainTokenSecret(cluster *corev1.StorageCluster, ownerRef *metav1.OwnerReference) error {
+	secret := &v1.Secret{}
+	err := c.k8sClient.Get(context.TODO(),
+		types.NamespacedName{
+			Name:      pxutil.PortworxServiceAccountTokenSecretName,
+			Namespace: cluster.Namespace,
+		}, secret)
+	if err != nil {
+		return fmt.Errorf("error getting the secret containing token for px container. %v", err)
+	}
+	needRefresh, err := isTokenRefreshRequired(secret)
+	if err != nil {
+		return err
+	}
+	if needRefresh {
+		newToken, err := generateToken(cluster)
+		if err != nil {
+			return err
+		}
+		curTime := time.Now()
+		tokenRefreshTimeBytes, err := curTime.Add(time.Duration(tokenExpirationSeconds / 2)).MarshalBinary()
+		if err != nil {
+			return fmt.Errorf("error marshalling current time to bytes. %v", err)
+		}
+		secret.Data[core.ServiceAccountTokenKey] = newToken
+		secret.Data[TokenRefreshTimeKey] = tokenRefreshTimeBytes
+		err = k8sutil.CreateOrUpdateSecret(c.k8sClient, secret, ownerRef)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func generateToken(cluster *corev1.StorageCluster) ([]byte, error) {
+	tokenRequest := &authv1.TokenRequest{
+		Spec: authv1.TokenRequestSpec{
+			Audiences:         []string{"px"},
+			ExpirationSeconds: &tokenExpirationSeconds,
+		},
+	}
+	tokenResp, err := coreops.Instance().CreateToken(pxutil.PortworxServiceAccountName(cluster), cluster.Namespace, tokenRequest)
+	if err != nil {
+		return nil, fmt.Errorf("error creating token from k8s. %v", err)
+	}
+	return []byte(tokenResp.Status.Token), nil
+}
+
+func isTokenRefreshRequired(secret *v1.Secret) (bool, error) {
+	expiryTime := time.Time{}
+	if err := expiryTime.UnmarshalBinary(secret.Data[TokenRefreshTimeKey]); err != nil {
+		return false, fmt.Errorf("error converting expiry time bytes to struct. %v", err)
+	}
+	if time.Now().After(expiryTime) {
+		return true, nil
+	}
+	return false, nil
 }
 
 func (c *portworxBasic) createPortworxKVDBService(
